lat5150drvmil/requirements.txt
# AI Framework - Dell Latitude 5450 MIL-SPEC
# Python 3.10+ required

# Deep Learning & AI
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
bitsandbytes>=0.41.0  # For 4-bit/8-bit quantization
peft>=0.5.0  # LoRA and PEFT
trl>=0.7.0  # DPO and PPO training
intel-extension-for-pytorch>=2.0.0  # Intel optimizations
openvino>=2023.0.0  # Intel NPU support

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: Experiment tracking

# RAG System & Vector Databases
sentence-transformers>=2.2.0
qdrant-client>=1.7.0  # Screenshot Intelligence System
faiss-cpu>=1.7.4  # Or faiss-gpu if using iGPU
chromadb>=0.4.0
langchain>=0.0.300
llama-index>=0.9.0

# AI APIs & Orchestration (MANDATORY)
google-generativeai>=0.8.0  # Gemini Pro - multimodal capabilities
openai>=1.0.0  # OpenAI API - structured outputs, GPT-4
duckduckgo-search>=6.3.0  # Web search integration
ollama>=0.3.0  # Local Ollama client
shodan>=1.31.0  # Threat intelligence

# Type-Safe AI with Pydantic (MANDATORY)
pydantic>=2.9.0  # Data validation and type safety
pydantic-ai>=0.0.13  # Type-safe agent framework
pydantic-settings>=2.5.0  # Settings management

# Screenshot Intelligence System
paddleocr>=2.7.0  # Primary OCR engine
paddlepaddle>=2.5.0  # PaddleOCR backend
pytesseract>=0.3.10  # Tesseract OCR wrapper (fallback)
Pillow>=10.0.0  # Image processing
watchdog>=3.0.0  # File system monitoring
psutil>=5.9.0  # System monitoring
mcp>=0.1.0  # Model Context Protocol

# Vector Search (AVX-512 module)
# C++ compilation dependencies
# Requires: g++, cmake, make

# Telegram Scrapers
telethon>=1.30.0
python-dotenv>=1.0.0

# OSINT Collection
# All OSINT collectors use stdlib only - no external dependencies!

# Web Scraping & APIs (optional for enhanced OSINT)
requests>=2.31.0  # Optional: Better HTTP client
beautifulsoup4>=4.12.0  # Optional: Web scraping

# Utilities
pyyaml>=6.0
toml>=0.10.2
click>=8.1.0  # CLI framework

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0  # Code formatting
flake8>=6.1.0  # Linting
mypy>=1.5.0  # Type checking

# ZFS utilities (system-level)
# Note: ZFS operations require zfs/zpool commands installed via:
#   sudo apt install zfsutils-linux

# DSMIL Framework (system-level)
# Note: DSMIL kernel module installed via:
#   sudo dpkg -i packaging/dsmil-complete_8.3.2-1.deb

# Development
ipython>=8.14.0
jupyter>=1.0.0
notebook>=7.0.0

# Web Framework & API
flask>=3.0.0
flask-cors>=4.0.0

# Optional: Production deployment
gunicorn>=21.2.0
uvicorn>=0.23.0  # For FastAPI if adding API endpoints
fastapi>=0.103.0

# Optional: Database (if needed for larger RAG deployments)
# psycopg2-binary>=2.9.7  # PostgreSQL
# redis>=5.0.0  # Caching
